# winalog-go/internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/collect": {
            "post": {
                "description": "在Windows上执行一键日志收集，收集系统信息、注册表、Prefetch、ShimCache等",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collect"
                ],
                "summary": "开始日志收集",
                "parameters": [
                    {
                        "description": "收集请求参数",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/api.LogCollectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.LogCollectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/collect/evtx2csv": {
            "post": {
                "description": "解析EVTX事件日志文件并导出为CSV格式",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collect"
                ],
                "summary": "将EVTX文件转换为CSV格式",
                "parameters": [
                    {
                        "description": "EVTX转换请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Evtx2CsvRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Evtx2CsvResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/collect/import": {
            "post": {
                "description": "从指定路径导入EVTX、ETL等日志文件到数据库",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collect"
                ],
                "summary": "导入日志文件",
                "parameters": [
                    {
                        "description": "导入请求，包含文件路径列表",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.LogImportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/collect/upload": {
            "post": {
                "description": "通过multipart/form-data上传日志文件到临时目录",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collect"
                ],
                "summary": "上传日志文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "要上传的文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success\": bool, \"filename\": string, \"path\": string, \"size\": int64",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/import/status": {
            "get": {
                "description": "Get the status of a log file import operation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "import"
                ],
                "summary": "Get import status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File path to check import status",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/types.ErrorCode"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.Evtx2CsvRequest": {
            "type": "object",
            "required": [
                "file_paths"
            ],
            "properties": {
                "calculate_hash": {
                    "type": "boolean"
                },
                "file_paths": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "include_xml": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "output_dir": {
                    "type": "string"
                }
            }
        },
        "api.Evtx2CsvResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "failed_files": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Result"
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "total_events": {
                    "type": "integer"
                },
                "total_files": {
                    "type": "integer"
                }
            }
        },
...
        "api.LogCollectOptions": {
            "type": "object",
            "properties": {
                "calculate_hash": {
                    "type": "boolean"
                },
                "compress": {
                    "type": "boolean"
                },
                "include_amcache": {
                    "type": "boolean"
                },
                "include_dlls": {
                    "type": "boolean"
                },
                "include_drivers": {
                    "type": "boolean"
                },
                "include_logs": {
                    "type": "boolean"
                },
                "include_network": {
                    "type": "boolean"
                },
                "include_prefetch": {
                    "type": "boolean"
                },
                "include_processes": {
                    "type": "boolean"
                },
                "include_registry": {
                    "type": "boolean"
                },
                "include_shimcache": {
                    "type": "boolean"
                },
                "include_system_info": {
                    "type": "boolean"
                },
                "include_tasks": {
                    "type": "boolean"
                },
                "include_userassist": {
                    "type": "boolean"
                },
                "include_users": {
                    "type": "boolean"
                },
                "include_usn_journal": {
                    "type": "boolean"
                },
                "output_path": {
                    "type": "string"
                },
                "workers": {
                    "type": "integer"
                }
            }
        },
        "api.LogCollectRequest": {
            "type": "object",
            "properties": {
                "formats": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "options": {
                    "$ref": "#/definitions/api.LogCollectOptions"
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "api.LogCollectResponse": {
            "type": "object",
            "properties": {
                "collected_items": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "duration": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "event_count": {
                    "type": "integer"
                },
                "failed_count": {
                    "type": "integer"
                },
                "hash": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "output_path": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "success_count": {
                    "type": "integer"
                }
            }
        },
        "api.LogImportRequest": {
            "type": "object",
            "required": [
                "files"
            ],
            "properties": {
                "alert_on_import": {
                    "type": "boolean"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "api.Result": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "event_count": {
                    "type": "integer"
                },
                "input_path": {
                    "type": "string"
                },
                "output_path": {
                    "type": "string"
                }
            }
        },
...
        "api.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "types.ErrorCode": {
            "type": "string",
            "enum": [
                "SUCCESS",
                "INTERNAL_ERROR",
                "INVALID_PARAM",
                "INVALID_REQUEST",
                "INVALID_QUERY",
                "NOT_FOUND",
                "EVENT_NOT_FOUND",
                "ALERT_NOT_FOUND",
                "ALERT_ALREADY_RESOLVED",
                "UNAUTHORIZED",
                "PARSE_FAILED",
                "IMPORT_FAILED",
                "FILE_NOT_FOUND",
                "FILE_LOCKED",
                "INVALID_FORMAT",
                "DB_ERROR",
                "DB_READ_ONLY",
                "RULE_INVALID",
                "RULE_NOT_FOUND",
                "RULE_DISABLED",
                "SEARCH_FAILED",
                "RESULT_TOO_LARGE",
                "HASH_MISMATCH",
                "SIGNATURE_INVALID",
                "NOT_SUPPORTED"
            ],
            "x-enum-varnames": [
                "ErrCodeSuccess",
                "ErrCodeInternalError",
                "ErrCodeInvalidParam",
                "ErrCodeInvalidRequest",
                "ErrCodeInvalidQuery",
                "ErrCodeNotFound",
                "ErrCodeEventNotFound",
                "ErrCodeAlertNotFound",
                "ErrCodeAlertAlreadyResolved",
                "ErrCodeUnauthorized",
                "ErrCodeParseFailed",
                "ErrCodeImportFailed",
                "ErrCodeFileNotFound",
                "ErrCodeFileLocked",
                "ErrCodeInvalidFormat",
                "ErrCodeDBError",
                "ErrCodeDBReadOnly",
                "ErrCodeRuleInvalid",
                "ErrCodeRuleNotFound",
                "ErrCodeRuleDisabled",
                "ErrCodeSearchFailed",
                "ErrCodeResultTooLarge",
                "ErrCodeHashMismatch",
                "ErrCodeSignatureInvalid",
                "ErrCodeNotSupported"
            ]
        },
...
    }
}